Added time and space complexity, minor changes
Need to put arr outside of function for memorization. 
Also, the question statement is wrong in input section. It says K should be less than N, but in picture K is 6 and N is just 4. 
Actually, K can go up to n*(n-1)/2 which is the same as C(n,2)